Minor details, such as int/double bug, turning off branches earlier in fit, reading mean in ST fits from file

+++ include/PhaseSpace/KKpipi_vs_Flavour_PhaseSpace.h
@@ -34,7 +34,8 @@ class KKpipi_vs_Flavour_PhaseSpace: public KKpipi_PhaseSpace {
     /**
      * Charge of the tag kaon
      */
-    double m_KaonCharge;
+    int m_KaonCharge;
+    //double m_KaonCharge; // Bug in BOSS code, will fix in the future
 };
 
 #endif
